+++ client/src/component/clusterinfo/action.js
@@ -3,7 +3,30 @@
 
 import type { ClusterMeta } from '../dashboard/action';
 import type { Event } from '../events/action';
+import type { ActionsSearchStore } from './store';
 
+
+export type Action = {
+  node_id: string,
+  action_id: string,
+  created_ts: string,
+  finished_ts: string,
+  kind: string,
+  state: string,
+  state_payload: ?Object,
+};
+
+export type ActionDetails = {
+  action: Action,
+  history: ActionHistory[],
+};
+
+export type ActionHistory = {
+  action_id: string,
+  timestamp: string,
+  state: string,
+  state_payload: ?Object
+};
 
 export type AgentDetails = {
   host: string,
@@ -11,33 +34,39 @@ export type AgentDetails = {
   version_checkout: ?string,
   version_number: ?string,
   version_taint: ?string,
-}
+};
 
 export type AgentStatus = {
   code: string,
   data?: string,
-}
+};
 
 export type ClusterDiscovery = {
   cluster_id: string,
   display_name: ?string,
   nodes: Array<string>,
-}
+};
 
 export type NodeInfo = {
   cluster_id: string,
   kind: string,
   node_id: string,
   version: string,
-}
+};
 
 
 /** Name actions for type checking. */
+export const CLUSTER_ACTIONS_SEARCH = 'CLUSTER_ACTIONS_SEARCH';
+export const CLUSTER_ACTIONS_SEARCH_FILTERS = 'CLUSTER_ACTIONS_SEARCH_FILTERS';
+export const CLUSTER_ACTIONS_SEARCH_STATE = 'CLUSTER_ACTIONS_SEARCH_STATE';
+export const CLUSTER_FETCH_ACTION = 'CLUSTER_FETCH_ACTION';
 export const CLUSTER_FETCH_AGENTS = 'CLUSTER_FETCH_AGENTS';
 export const CLUSTER_FETCH_DISCOVERY = 'CLUSTER_FETCH_DISCOVERY';
 export const CLUSTER_FETCH_EVENTS = 'CLUSTER_FETCH_EVENTS';
 export const CLUSTER_FETCH_META = 'CLUSTER_FETCH_META';
 export const CLUSTER_FETCH_NODES = 'CLUSTER_FETCH_NODES';
+export const CLUSTER_STORE_ACTION = 'CLUSTER_STORE_ACTION';
+export const CLUSTER_STORE_ACTIONS = 'CLUSTER_STORE_ACTIONS';
 export const CLUSTER_STORE_AGENTS = 'CLUSTER_STORE_AGENTS';
 export const CLUSTER_STORE_DISCOVERY = 'CLUSTER_STORE_DISCOVERY';
 export const CLUSTER_STORE_EVENTS = 'CLUSTER_STORE_EVENTS';
@@ -46,10 +75,35 @@ export const CLUSTER_STORE_NODES = 'CLUSTER_STORE_NODES';
 
 
 /** Type enum of all possible fetch actions. */
+export type ClusterActionsSearchAction = {
+  +type: typeof CLUSTER_ACTIONS_SEARCH,
+  +cluster_id: string,
+  +search: ActionsSearchStore,
+};
+
+export type ClusterActionsSearchFiltersAction = {
+  +type: typeof CLUSTER_ACTIONS_SEARCH_FILTERS,
+  +cluster_id: string,
+  +search: ActionsSearchStore,
+};
+
+export type ClusterActionsSearchStateAction = {
+  +type: typeof CLUSTER_ACTIONS_SEARCH_STATE,
+  +cluster_id: string,
+  +state: boolean,
+};
+
+export type ClusterFetchActionAction = {
+  +type: typeof CLUSTER_FETCH_ACTION,
+  +cluster_id: string,
+  +action_id: string,
+};
+
 export type ClusterFetchAgentsAction = {
   +type: typeof CLUSTER_FETCH_AGENTS,
   +cluster_id: string,
 };
+
 export type ClusterFetchDiscoveryAction = {
   +type: typeof CLUSTER_FETCH_DISCOVERY,
   +cluster_id: string,
@@ -67,6 +121,18 @@ export type ClusterFetchMetaAction = {
 
 export type ClusterFetchNodesAction = {
   +type: typeof CLUSTER_FETCH_NODES,
+  +cluster_id: string,
+};
+
+export type ClusterStoreActionAction = {
+  +type: typeof CLUSTER_STORE_ACTION,
+  +action: ActionDetails,
+  +cluster_id: string,
+};
+
+export type ClusterStoreActionsAction = {
+  +type: typeof CLUSTER_STORE_ACTIONS,
+  +actions: Action[],
   +cluster_id: string,
 };
 
@@ -99,11 +165,17 @@ export type ClusterStoreNodesAction = {
 };
 
 export type ClusterInfoAction =
+  ClusterActionsSearchAction |
+  ClusterActionsSearchFiltersAction |
+  ClusterActionsSearchStateAction |
+  ClusterFetchActionAction |
   ClusterFetchAgentsAction |
   ClusterFetchDiscoveryAction |
   ClusterFetchEventsAction |
   ClusterFetchMetaAction |
   ClusterFetchNodesAction |
+  ClusterStoreActionAction |
+  ClusterStoreActionsAction |
   ClusterStoreAgentsAction |
   ClusterStoreDiscoveryAction |
   ClusterStoreEventsAction |
@@ -111,6 +183,14 @@ export type ClusterInfoAction =
   ClusterStoreNodesAction |
   {type: 'FLOW_CATCH_ALL'};
 
+
+export function fetchAction(cluster: string, action: string): ClusterFetchActionAction {
+  return {
+    type: CLUSTER_FETCH_ACTION,
+    cluster_id: cluster,
+    action_id: action,
+  };
+}
 
 export function fetchAgents(cluster: string): ClusterFetchAgentsAction {
   return {
@@ -144,5 +224,27 @@ export function fetchNodes(cluster: string): ClusterFetchNodesAction {
   return {
     type: CLUSTER_FETCH_NODES,
     cluster_id: cluster,
+  };
+}
+
+export function saveActionsFilters(
+  cluster: string,
+  search: ActionsSearchStore
+): ClusterActionsSearchFiltersAction {
+  return {
+    type: CLUSTER_ACTIONS_SEARCH_FILTERS,
+    cluster_id: cluster,
+    search,
+  };
+}
+
+export function searchActions(
+  cluster: string,
+  search: ActionsSearchStore
+): ClusterActionsSearchAction {
+  return {
+    type: CLUSTER_ACTIONS_SEARCH,
+    cluster_id: cluster,
+    search,
   };
 }