+++ client/src/component/clusterinfo/saga.js
@@ -3,25 +3,35 @@
 
 import { all } from 'redux-saga/effects';
 import { call, put } from 'redux-saga/effects';
+import { delay } from 'redux-saga/effects';
 import { takeEvery } from 'redux-saga/effects';
 
+import { CLUSTER_ACTIONS_SEARCH } from './action';
+import { CLUSTER_ACTIONS_SEARCH_STATE } from './action';
+import { CLUSTER_FETCH_ACTION } from './action';
 import { CLUSTER_FETCH_AGENTS } from './action';
 import { CLUSTER_FETCH_DISCOVERY } from './action';
 import { CLUSTER_FETCH_EVENTS } from './action';
 import { CLUSTER_FETCH_META } from './action';
 import { CLUSTER_FETCH_NODES } from './action';
+import { CLUSTER_STORE_ACTION } from './action';
+import { CLUSTER_STORE_ACTIONS } from './action';
 import { CLUSTER_STORE_AGENTS } from './action';
 import { CLUSTER_STORE_DISCOVERY } from './action';
 import { CLUSTER_STORE_EVENTS } from './action';
 import { CLUSTER_STORE_META } from './action';
 import { CLUSTER_STORE_NODES } from './action';
 
+import { fetchAction as fetchActionApi } from './api';
+import { fetchActions as fetchActionsApi } from './api';
 import { fetchAgents as fetchAgentsApi } from './api';
 import { fetchDiscovery as fetchDiscoveryApi } from './api';
 import { fetchEvents as fetchEventsApi } from './api';
 import { fetchMeta as fetchMetaApi } from './api';
 import { fetchNodes as fetchNodesApi } from './api';
 
+import type { ClusterActionsSearchAction } from './action';
+import type { ClusterFetchActionAction } from './action';
 import type { ClusterFetchAgentsAction } from './action';
 import type { ClusterFetchDiscoveryAction } from './action';
 import type { ClusterFetchEventsAction } from './action';
@@ -35,6 +45,18 @@ import type { ClusterFetchNodesAction } from './action';
 export function* fetchAgents(action: ClusterFetchAgentsAction): any {
   let agents = yield call(fetchAgentsApi, action.cluster_id);
   yield put({type: CLUSTER_STORE_AGENTS, agents: agents, cluster_id: action.cluster_id});
+}
+
+/**
+ * Fetches a cluster's action and stores it in redux.
+ */
+export function* fetchAction(action: ClusterFetchActionAction): any {
+  let details = yield call(fetchActionApi, action.cluster_id, action.action_id);
+  yield put({
+    type: CLUSTER_STORE_ACTION,
+    cluster_id: action.cluster_id,
+    action: details,
+  });
 }
 
 /**
@@ -69,9 +91,24 @@ export function* fetchNodes(action: ClusterFetchNodesAction): any {
   yield put({type: CLUSTER_STORE_NODES, cluster_id: action.cluster_id, nodes: nodes});
 }
 
+/** Execute an actions search. */
+export function* searchActions(action: ClusterActionsSearchAction): any {
+  yield put({type: CLUSTER_ACTIONS_SEARCH_STATE, cluster_id: action.cluster_id, state: true});
+  try {
+    let actions = yield call(fetchActionsApi, action.cluster_id, action.search);
+    yield put({type: CLUSTER_STORE_ACTIONS, cluster_id: action.cluster_id, actions})
+  } catch (error) {
+    console.error("Failed to search for actions", error);
+  } finally {
+    yield put({type: CLUSTER_ACTIONS_SEARCH_STATE, cluster_id: action.cluster_id, state: false});
+  }
+}
+
 /** Main saga for the datafetch component. */
 export function* saga(): any {
   yield all([
+    takeEvery(CLUSTER_ACTIONS_SEARCH, searchActions),
+    takeEvery(CLUSTER_FETCH_ACTION, fetchAction),
     takeEvery(CLUSTER_FETCH_AGENTS, fetchAgents),
     takeEvery(CLUSTER_FETCH_DISCOVERY, fetchDiscovery),
     takeEvery(CLUSTER_FETCH_EVENTS, fetchEvents),