+++ client/src/component/clusterinfo/store.js
@@ -3,15 +3,25 @@
 
 import type { ClusterMeta } from '../dashboard/action';
 
+import { CLUSTER_ACTIONS_SEARCH_FILTERS } from './action';
+import { CLUSTER_ACTIONS_SEARCH_STATE } from './action';
+import { CLUSTER_STORE_ACTION } from './action';
+import { CLUSTER_STORE_ACTIONS } from './action';
 import { CLUSTER_STORE_AGENTS } from './action';
 import { CLUSTER_STORE_DISCOVERY } from './action';
 import { CLUSTER_STORE_EVENTS } from './action';
 import { CLUSTER_STORE_META } from './action';
 import { CLUSTER_STORE_NODES } from './action';
 
+import type { Action } from './action';
+import type { ActionDetails } from './action';
 import type { AgentDetails } from './action';
+import type { ClusterActionsSearchFiltersAction } from './action';
+import type { ClusterActionsSearchStateAction } from './action';
 import type { ClusterDiscovery } from './action';
 import type { ClusterInfoAction } from './action';
+import type { ClusterStoreActionAction } from './action';
+import type { ClusterStoreActionsAction } from './action';
 import type { ClusterStoreAgentsAction } from './action';
 import type { ClusterStoreDiscoveryAction } from './action';
 import type { ClusterStoreEventsAction } from './action';
@@ -21,7 +31,23 @@ import type { Event } from '../events/action';
 import type { NodeInfo } from './action';
 
 
+export type ActionsSearchStore = {
+  action_kind: string,
+  action_state: string,
+  from: Date,
+  node_id: string,
+  searching: boolean,
+  until: Date,
+};
+
+export type ActionsStore = {
+  actions: Action[],
+  details: ?ActionDetails,
+  search: ActionsSearchStore,
+}
+
 export type ClusterInfoStore = {
+  actions: {[string]: ActionsStore},
   agents: {[string]: Array<AgentDetails>},
   discovery: {[string]: ClusterDiscovery},
   events: {[string]: Array<Event>},
@@ -31,6 +57,7 @@ export type ClusterInfoStore = {
 
 
 export const defaultState: ClusterInfoStore = {
+  actions: {},
   agents: {},
   discovery: {},
   events: {},
@@ -39,8 +66,29 @@ export const defaultState: ClusterInfoStore = {
 };
 
 
+export function defaultActionsStore() {
+  let now = new Date();
+  let from = new Date(now.getFullYear(), now.getMonth(), now.getDate() - 1);
+  return {
+    actions: [],
+    details: null,
+    search: {
+      action_kind: '',
+      action_state: '',
+      from,
+      node_id: '',
+      searching: false,
+      until: now,
+    },
+  };
+}
+
 export function reducer(state: ClusterInfoStore = defaultState, action: ClusterInfoAction) {
   switch (action.type) {
+    case CLUSTER_ACTIONS_SEARCH_FILTERS: return storeActionsFilters(state, action);
+    case CLUSTER_ACTIONS_SEARCH_STATE: return storeActionsSearching(state, action);
+    case CLUSTER_STORE_ACTION: return storeAction(state, action);
+    case CLUSTER_STORE_ACTIONS: return storeActions(state, action);
     case CLUSTER_STORE_AGENTS: return storeAgents(state, action);
     case CLUSTER_STORE_DISCOVERY: return storeDiscovery(state, action);
     case CLUSTER_STORE_EVENTS: return storeEvents(state, action);
@@ -50,6 +98,61 @@ export function reducer(state: ClusterInfoStore = defaultState, action: ClusterI
     default:
       return state;
   }
+}
+
+
+function storeActionsFilters(state: ClusterInfoStore, action: ClusterActionsSearchFiltersAction) {
+  let newState: ClusterInfoStore = {
+    ...state,
+    actions: {...state.actions},
+  };
+  let actions = newState.actions[action.cluster_id] || defaultActionsStore();
+  actions = {
+    ...actions,
+    search: action.search,
+  };
+  newState.actions[action.cluster_id] = actions;
+  return newState;
+}
+
+
+function storeActionsSearching(state: ClusterInfoStore, action: ClusterActionsSearchStateAction) {
+  let newState: ClusterInfoStore = {
+    ...state,
+    actions: {...state.actions},
+  };
+  let actions = newState.actions[action.cluster_id] || defaultActionsStore();
+  actions = {...actions};
+  actions.search = {...actions.search};
+  actions.search.searching = action.state;
+  newState.actions[action.cluster_id] = actions;
+  return newState;
+}
+
+
+function storeAction(state: ClusterInfoStore, action: ClusterStoreActionAction) {
+  let newState: ClusterInfoStore = {
+    ...state,
+    actions: {...state.actions},
+  };
+  let actions = newState.actions[action.cluster_id] || defaultActionsStore();
+  actions = {...actions};
+  actions.details = action.action;
+  newState.actions[action.cluster_id] = actions;
+  return newState;
+}
+
+
+function storeActions(state: ClusterInfoStore, action: ClusterStoreActionsAction) {
+  let newState: ClusterInfoStore = {
+    ...state,
+    actions: {...state.actions},
+  };
+  let actions = newState.actions[action.cluster_id] || defaultActionsStore();
+  actions = {...actions};
+  actions.actions = action.actions;
+  newState.actions[action.cluster_id] = actions;
+  return newState;
 }