+++ client/src/component/dashboard/__tests__/events.js
@@ -38,12 +38,12 @@ describe('Dashboard', () => {
 
     test('renders with events', () => {
       let events = [{
-        data: {"SOME": "DATA"},
         event: "AGENT_DOWN",
+        payload: {"SOME": "DATA"},
         timestamp: "2018-04-28T17:57:24.540480643Z"
       }, {
-        data: {"SOME": "DATA"},
         event: "AGENT_RECOVER",
+        payload: {"SOME": "DATA"},
         timestamp: "2018-04-28T17:50:15.170187929Z"
       }];
       let store = mockStore({