+++ server/index.js
@@ -12,9 +12,19 @@ const STATIC_ROOT = path.join(DIST_ROOT, 'static');
 
 
 const app = express();
+function notFound(_, res) {
+  res.status(404).json({error: 'Not Found'});
+}
 
+// Handle API and assets requests.
 app.use('/api', proxy);
 app.use('/static', express.static(STATIC_ROOT));
+
+// Fail unhandled API and assets requests.
+app.use('/api', notFound);
+app.use('/static', notFound);
+
+// For everything else serve the index page which will use the URL to route itself.
 app.use((_, res) => res.sendFile(path.join(DIST_ROOT, 'index.html')));