+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Cluster specific actions tab.
+- Render action details and history.
+
+### Fixed
+- Reflect events schema changes.
 
 ## [0.3.1] - 2019-07-15
 ### Added